Begin separating tasks into individual files
This first step moves the flash compilation into it's own taskfile
